pyproject.toml
[project]
name = "spark-kafka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python==2.0.2",
    "pyspark==3.5.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "ruff>=0.12.1",
    "schedule==1.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.1",
]
